created hook for running analysis
