+++ apps/serpeo/src/components/analyze-seo-input/use-analyze-seo-input.tsx
@@ -1,0 +1,32 @@
+import { useNavigate } from "@tanstack/react-router";
+import { useEffect } from "react";
+import { commands, events } from "../../generated/bindings";
+
+export const useAnalyzeSeoInput = () => {
+  const navigate = useNavigate();
+
+  const analyzeSeo = async (url: string) => {
+    try {
+      // setLoading(true);
+      // setResult(RESET);
+      const _analysis = await commands.analyzeUrlSeo(url);
+    } catch (error) {
+      console.error("Error analyzing SEO:", error);
+    }
+  };
+
+  useEffect(() => {
+    events.siteRunIdSet.listen((e) => {
+      console.log("siteRunIdSet");
+      console.log(e.payload);
+      navigate({
+        to: "/analysis/$site-run-id",
+        params: { "site-run-id": `${e.payload.site_run_id}` },
+      });
+    });
+  }, [navigate]);
+
+  return {
+    analyzeSeo,
+  };
+};
